% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualhalfing.R
\name{annualhalfing_LL}
\alias{annualhalfing_LL}
\title{Likelihood maximization for annual halfing customer retention model}
\usage{
annualhalfing_LL(grid, dta)
}
\arguments{
\item{grid}{model parameters}

\item{dta}{dataset}
}
\value{
Annual halfing Likelihood in optimization routine
}
\description{
Likelihood maximization for annual halfing customer retention model
}
