% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigT_expand_via_apply.R
\name{bigT_expand_via_apply}
\alias{bigT_expand_via_apply}
\title{bigT_expand_via_apply}
\usage{
bigT_expand_via_apply(x)
}
\arguments{
\item{x}{vector containing bigT, cancel and count}
}
\description{
bigT_expand_via_apply
}
\examples{
x <- c(3, 1, 5)
bigT_expand_via_apply(x)
}
