% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_generate_manual.R
\name{book_generate_manual}
\alias{book_generate_manual}
\title{generates the PDF manual for the R CADF package}
\usage{
book_generate_manual()
}
\value{
PDF documentation
}
\description{
`book_generate_manual` inputs are search term
}
\details{
Description here
}
