% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_print_sources.R
\name{book_print_sources}
\alias{book_print_sources}
\title{print source list based on search term}
\usage{
book_print_sources(searchterm)
}
\description{
`book_print_sources` inputs are search term
}
\details{
Description here
}
