% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_SRM.R
\name{ca_SRM}
\alias{ca_SRM}
\title{ca_SRM}
\usage{
ca_SRM(df_logistic)
}
\arguments{
\item{df_logistic}{data frame containing the data for logistic regression}
}
\description{
ca_SRM
}
\examples{
customertype1 <- c(3, 1, 5)
customertype2 <- c(12, 0, 3)
cust1 <- bigT_expand_via_apply(customertype1)
cust2 <- bigT_expand_via_apply(customertype2)
df_logistic <- rbind(cust1, cust2)
model <- ca_SRM(df_logistic)
}
