% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CADF_to_annualhalfing_data.R
\name{CADF_to_annualhalfing_data}
\alias{CADF_to_annualhalfing_data}
\title{Convert CADF dataset into annualhalfing model dataset}
\usage{
CADF_to_annualhalfing_data(cadf.data)
}
\arguments{
\item{cadf.data}{CADF dataset}
}
\description{
Converts CADF output to dataset for annual halfing model
}
