% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CADF_to_btyd_pareto_nbd.R
\name{CADF_to_btyd_pareto_nbd}
\alias{CADF_to_btyd_pareto_nbd}
\title{CADF to btyd pareto nbd model}
\usage{
CADF_to_btyd_pareto_nbd(cadf.data)
}
\arguments{
\item{cadf.data}{CADF-formatted dataset}
}
\description{
Converts a CADF dataset to a dataset for btyd pareto nbd modeling
}
