% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CADF_to_logistic_regression.R
\name{CADF_to_logistic_regression}
\alias{CADF_to_logistic_regression}
\title{CADF to logistic regression}
\usage{
CADF_to_logistic_regression(CADF)
}
\arguments{
\item{CADF}{CADF-formatted dataset}
}
\description{
Convert a CADF dataset to a dataset for logistic regression
}
