% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CADF_to_migration_model.R
\name{CADF_to_migration_model}
\alias{CADF_to_migration_model}
\title{CADF_to_migration_model converts CADF data to migration model data}
\usage{
CADF_to_migration_model(cadf.data, maxT = 5)
}
\arguments{
\item{cadf.data}{Data in R list format processed by CADF functions}

\item{maxT}{If time is greater than maxT it will be converted into a + category}
}
\description{
Builds transition matrix for a migration model.
T is the maximum time cutoff which defaults to 5.   
The output will be a transition matrix.
}
\examples{
tmatrix <- CADF_to_migration_model(cadf.data.sample)
}
