% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CADF_to_nth_purchase.R
\name{CADF_to_nth_purchase}
\alias{CADF_to_nth_purchase}
\title{CADF_to_nth_purchase}
\usage{
CADF_to_nth_purchase(cadf.data, n)
}
\arguments{
\item{cadf.data}{Data in R list format processed by CADF functions}

\item{n}{the nth purchase you want to analyze}
}
\description{
CADF_to_nth_purchase
}
