% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CADF_to_nth_purchase_allrows.R
\name{CADF_to_nth_purchase_allrows}
\alias{CADF_to_nth_purchase_allrows}
\title{CADF_to_nth_purchase_allrows inputs CADF data and the desired purchase number that you want to count the nth result of.}
\usage{
CADF_to_nth_purchase_allrows(cadf.data, n)
}
\arguments{
\item{cadf.data}{Data in R list format processed by CADF functions}

\item{n}{the nth purchase}
}
\description{
CADF_to_nth_purchase_allrows inputs CADF data and the desired purchase number that you want to count the nth result of.
}
