% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.recency.matrix.R
\name{create.recency.string}
\alias{create.recency.string}
\title{create_recency_string}
\usage{
create.recency.string(x)
}
\arguments{
\item{x}{vector of zeros and ones}
}
\description{
Tracks cumulative recency
}
\examples{
head(cadf.data.sample)
}
