% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_CustomerModelingMatrix.R
\name{f_CustomerModelingMatrix}
\alias{f_CustomerModelingMatrix}
\title{For each customer, return a modeling matrix that is utilized for logistic regression}
\usage{
f_CustomerModelingMatrix(cancellation_time)
}
\arguments{
\item{cancellation_time}{= cancellation time}
}
\description{
`f_CustomerModelingMatrix` inputs are cancellation_time.
}
\details{
Description here
}
\examples{
f_CustomerModelingMatrix(10)
}
