% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_CustomerSurvivalModelingMatrix.R
\name{f_CustomerSurvivalModelingMatrix}
\alias{f_CustomerSurvivalModelingMatrix}
\title{For each customer, return a survival modeling matrix that is utilized for survival analysis}
\usage{
f_CustomerSurvivalModelingMatrix(cancellation_time)
}
\arguments{
\item{cancellation_time}{cancellation time}
}
\description{
`f_CustomerSurvivalModelingMatrix` inputs are T.
}
\details{
Description here
}
\examples{
f_CustomerSurvivalModelingMatrix(10)
}
