% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_intMonths.R
\name{f_intMonths}
\alias{f_intMonths}
\title{Compute the months between two purchase dates}
\usage{
f_intMonths(a, b)
}
\arguments{
\item{a}{starting date}

\item{b}{ending date

Description here}
}
\description{
Compute the months between two purchase dates
}
