% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_from_ps.R
\name{frequency_from_ps}
\alias{frequency_from_ps}
\title{Purchase string to frequency count}
\usage{
frequency_from_ps(x)
}
\arguments{
\item{x}{rle object}
}
\description{
Purchase string to frequency count
}
