% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_from_rle.R
\name{frequency_from_rle}
\alias{frequency_from_rle}
\title{RLE object to frequency count}
\usage{
frequency_from_rle(x)
}
\arguments{
\item{x}{rle object}
}
\description{
RLE object to frequency count
}
\examples{
# example code
x <- c(1,1,0,1,0,0,1,0,0,0)
x.rle <- rle(x)
frequency_from_rle(x.rle)
}
