% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadf.R
\docType{data}
\name{gammagamma}
\alias{gammagamma}
\title{Gamma gamma spend model data}
\format{
data frame with 2,357 rows and 6 columns
}
\description{
Gamma gamma spend model data
}
\keyword{gamma}
\keyword{model}
\keyword{spend}
