% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_date_template.R
\name{generate_date_template}
\alias{generate_date_template}
\title{generate_date_template}
\usage{
generate_date_template()
}
\description{
generate_date_template
}
\examples{
dates <- generate_date_template()
}
