% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_to_CADF.R
\name{id_to_CADF}
\alias{id_to_CADF}
\title{Convert to CADF for a single customer id}
\usage{
id_to_CADF(data, today.study.cutoff)
}
\description{
`id_to_CADF` inputs is coming from a lapply operation on a split customer dataset.  If variable a is the split customer dataset then a$`1` is customer with ID 1
}
\details{
Description here
}
