% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_sample_customer_matrix.R
\name{ld_sample_customer_matrix}
\alias{ld_sample_customer_matrix}
\title{LD functions are utilized for learning and diagnostic use.}
\usage{
ld_sample_customer_matrix(numCustomers, maxT, purchaseAtT0 = TRUE)
}
\arguments{
\item{numCustomers}{number of customers to simulate}

\item{maxT}{number of timeperiods}

\item{purchaseAtT0}{by default sets first column of matrix to 1}
}
\description{
LD functions are utilized for learning and diagnostic use.
}
