% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnualHalfingLikelihoodOptimization.R
\name{modeling.annualhalfing.likelihood}
\alias{modeling.annualhalfing.likelihood}
\title{Likelihood function for annual halfing model}
\usage{
modeling.annualhalfing.likelihood(grid2, rec, freq, targetBuy)
}
\arguments{
\item{grid2}{Modeling parameters}

\item{rec}{recency}

\item{freq}{frequency}

\item{targetBuy}{indicator if purchase was made in holdout period}
}
\description{
Likelihood function for annual halfing model
}
