% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.LL.gamma_spend.R
\name{modeling.LL.gamma_spend}
\alias{modeling.LL.gamma_spend}
\title{LL function for the gamma gamma spend model}
\usage{
modeling.LL.gamma_spend(p, q, gamma, y = data)
}
\arguments{
\item{p}{p}

\item{q}{q}

\item{gamma}{gamma}

\item{y}{data}
}
\description{
LL function for the gamma gamma spend model
}
