% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_gamma.R
\name{pdf_gamma}
\alias{pdf_gamma}
\title{PDF probability function for gamma distribution}
\usage{
pdf_gamma(x, r, a)
}
\arguments{
\item{x}{between 0 and 1 for pdf}

\item{r}{shape parameter}

\item{a}{scale parameter}
}
\description{
PDF probability function for gamma distribution
}
