% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_gamma2.R
\name{pdf_gamma2}
\alias{pdf_gamma2}
\title{Probability density function for gamma distribution}
\usage{
pdf_gamma2(x, shape, scale)
}
\arguments{
\item{x}{x}

\item{shape}{shape parameter}

\item{scale}{scale parameter}
}
\description{
Probability density function for gamma distribution
}
