% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_to_T_custom.R
\name{ps_to_T_custom}
\alias{ps_to_T_custom}
\title{Calculates T from a purchase string.  Custom.}
\usage{
ps_to_T_custom(ps, skips = 2)
}
\arguments{
\item{ps}{Purchase string.}

\item{skips}{Number of non purchase periods that the customer is still considered a customer for.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Calculates T from a purchase string.  Custom.
}
