% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_to_T_strict_quitter.R
\name{ps_to_T_strict_quitter}
\alias{ps_to_T_strict_quitter}
\title{Calculates T from a purchase string}
\usage{
ps_to_T_strict_quitter(ps)
}
\arguments{
\item{ps}{Purchase string.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Calculates T from a purchase string
}
