% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_to_T_strict_stayer.R
\name{ps_to_T_strict_stayer}
\alias{ps_to_T_strict_stayer}
\title{Calculates T from a purchase string under the "strict stayer" assumption.}
\usage{
ps_to_T_strict_stayer(ps)
}
\arguments{
\item{ps}{Purchase string.}
}
\value{
The numeric value for T, which is the position of the last 1 in the purchase string
}
\description{
Calculates T from a purchase string under the "strict stayer" assumption.
}
