% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmatrix_to_psstring.R
\name{psmatrix_to_psstring}
\alias{psmatrix_to_psstring}
\title{psmatrix_to_psstring}
\usage{
psmatrix_to_psstring(psmatrix)
}
\arguments{
\item{psmatrix}{purchase string of 1's and 0's in matrix format}
}
\description{
psmatrix_to_psstring
}
\examples{
cadf.data.sample[[4]]$purchase_string_as_matrix
}
