% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmatrix_to_recency_at_time_of_matrix.R
\name{psmatrix_to_recency_attimeof_matrix}
\alias{psmatrix_to_recency_attimeof_matrix}
\title{accepts a psmatrix
converts 1/0 purchase strings to recency at timeof}
\usage{
psmatrix_to_recency_attimeof_matrix(psmatrix)
}
\description{
accepts a psmatrix
converts 1/0 purchase strings to recency at timeof
}
