% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_purchase_string.R
\name{qc_purchase_string}
\alias{qc_purchase_string}
\title{Accepts purchase string as an arguement and returns Y/N
based on whether or not the purchase string is valid
for the customer analytics data format (CADF)}
\usage{
qc_purchase_string(ps)
}
\description{
Accepts purchase string as an arguement and returns Y/N
based on whether or not the purchase string is valid
for the customer analytics data format (CADF)
}
