% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_transactional_data.R
\name{qc_transactional_data}
\alias{qc_transactional_data}
\title{The customer analytics data format (CADF) relays heavily on correct input data.
Transactional data must:
1.) be a data frame with two columns
2.) Column one is the customer id
3.) Column 2 is the transaction date.  Column 2 must be formatted as a date object in R.}
\usage{
qc_transactional_data(x)
}
\arguments{
\item{x}{R dataframe representing ..}
}
\value{
A number representing whether it passes or not.
}
\description{
The customer analytics data format (CADF) relays heavily on correct input data.
Transactional data must:
1.) be a data frame with two columns
2.) Column one is the customer id
3.) Column 2 is the transaction date.  Column 2 must be formatted as a date object in R.
}
