% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recency_binning.R
\name{recency_binning}
\alias{recency_binning}
\title{Create recency bins
Create recency bins}
\usage{
recency_binning(x, max.recency)
}
\value{
string representing categorical value
}
\description{
Create recency bins
Create recency bins
}
\details{
Create recency bins
}
