% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recency_from_rle.R
\name{recency_from_rle}
\alias{recency_from_rle}
\title{Input must be attr(v3, "class")
The following function takes an R rle object as input.  Rle is base R functionality and information can be found by ?rle.
Rle can be ran on a vector of 1/0 to determine the number of time periods spent in each state.  This function reverses the information returned from rle.  If the streak ends in 0 (x$values = 0) then the value (x$lengths) associated with zero is returned.  Otherwise, 0 is returned.  The streak of zeros tells us the recency.  Larger numbers equal more latency between purchases.}
\usage{
recency_from_rle(x)
}
\description{
Input must be attr(v3, "class")
The following function takes an R rle object as input.  Rle is base R functionality and information can be found by ?rle.
Rle can be ran on a vector of 1/0 to determine the number of time periods spent in each state.  This function reverses the information returned from rle.  If the streak ends in 0 (x$values = 0) then the value (x$lengths) associated with zero is returned.  Otherwise, 0 is returned.  The streak of zeros tells us the recency.  Larger numbers equal more latency between purchases.
}
