% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.migration.R
\name{simple_migration}
\alias{simple_migration}
\title{Simple Migration}
\usage{
simple_migration(num.customers, pct.buy.buy, pct.nobuy.buy, n.periods)
}
\arguments{
\item{num.customers}{Number of customers for the simulation.}

\item{pct.buy.buy}{percentage of customers that buy in the nxt period}

\item{pct.nobuy.buy}{percentage of non buyers that convert over to buyers}

\item{n.periods}{number of periods}
}
\description{
Function used for simulation and scenario planning
}
\examples{
simple_migration(200, .80, .20, 12)

}
