% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_transaction_file_to_CADF.R
\name{split.transaction.file_to_CADF}
\alias{split.transaction.file_to_CADF}
\title{Create a CADF dataset from a dataframe}
\usage{
\method{split}{transaction.file_to_CADF}(data, today.study.cutoff)
}
\arguments{
\item{data}{data frame for a single customer id}

\item{today.study.cutoff}{separate analysis and holdout data}
}
\description{
Create a CADF dataset from a dataframe
}
