% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadf.R
\docType{data}
\name{srm_data}
\alias{srm_data}
\title{#' Simple retention model data}
\format{
A data frame with 5828 rows and two columns
\describe{
  \item{bigT}{Time period}
  \item{cancel}{Whether or not there was a cancellation in the time period}
  ...
}
}
\description{
#' Simple retention model data
}
