% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadf.R
\name{transactions.merged}
\alias{transactions.merged}
\title{#' Transaction data}
\format{
A data frame with 67,944 rows and 4 columns
\describe{
  \item{ID}{Customer ID}
  \item{PURCHASE_DATE}{Purchase date}
  \item{NUM_ITEMS}{Number of items purchased}
  \item{TOTAL}{Total transaction amount}
  ...
}
}
\description{
#' Transaction data
}
