% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{transitions}
\alias{transitions}
\title{Calculate transition periods between two timeperiods}
\usage{
transitions(timeperiod0, timeperiod1, buyvar = "Y", nobuyvar = "N")
}
\arguments{
\item{timeperiod0}{Column representing the 'from' side of the transition probability}

\item{timeperiod1}{Column representing the 'to' side of the transition probability}

\item{buyvar}{field value that represents a buy, defaults to Y}

\item{nobuyvar}{field value that represents not buy, defaults to N}
}
\value{
2 x 2 transaction matrix
}
\description{
Calculate transition periods between two timeperiods
}
\examples{
timeperiod0 <- c("Y", "Y", "Y", "Y", "Y")
timeperiod1 <- c("N", "Y", "N", "Y", "N")
transitions(timeperiod0, timeperiod1)
}
