% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CADF_to_migration_model.R
\name{CADF_to_migration_model}
\alias{CADF_to_migration_model}
\title{Builds dataset for a migration model from CADF customer profiles.
T is the maximum time cutoff which defaults to 3.   The output matrix will be a t by t transition matrix}
\usage{
CADF_to_migration_model(dta)
}
\arguments{
\item{T}{Number of timeperiods (size of matrix)}

\item{cadf_data}{Data in R list format processed by CADF functions}
}
\description{
`ds.migration.model` generates a transaction matrix for a migration model
}
\details{
purchase strings are input as a 1 x 1 character matrix
Y represents buy; N represents no buy
}
