% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.purchase.string.R
\name{create.purchase.string}
\alias{create.purchase.string}
\title{Function called during Customer$new() (the Customer R6 class) to create purchase string for the customer.}
\usage{
create.purchase.string(x, id.column, date.column, return.mode = "")
}
\arguments{
\item{x}{Transactional data associated with customer id.}

\item{id.column}{Description goes here.}

\item{date.column.}{Description goes here.}
}
\value{
purchase string in 0/1 format.  Returned as string.
}
\description{
Function called during Customer$new() (the Customer R6 class) to create purchase string for the customer.
}
