% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_CustomerModelingMatrix.R
\name{f_CustomerModelingMatrix}
\alias{f_CustomerModelingMatrix}
\title{For each customer, return a modeling matrix that is utilized for logistic regression}
\usage{
f_CustomerModelingMatrix(T)
}
\description{
`f_CustomerModelingMatrix` inputs are T.
}
\details{
Description here
}
\section{Fields}{

\describe{
\item{\code{T}}{= cancellation time}
}}

\examples{
f_CustomerModelingMatrix(10)
}
