% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_intMonths.R
\name{f_intMonths}
\alias{f_intMonths}
\title{Compute the months between two purchase dates}
\usage{
f_intMonths(a, b)
}
\description{
`f_intMonths` inputs are customer id and transactional data.
}
\details{
Description here
}
